requirements.txt
python-dotenv 
requests
gunicorn
google-cloud-texttospeech 
google-auth
fastapi 
uvicorn 